% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_and_merge_indoor.R
\name{clean_and_merge_indoor}
\alias{clean_and_merge_indoor}
\title{clean_and_merge_indoor Title}
\usage{
clean_and_merge_indoor(
  raw_data_file,
  sh_name = "Data",
  eat = "Indian/Antananarivo"
)
}
\value{
tibble
}
\description{
clean_and_merge_indoor Title
}
