% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_inputs.R
\name{link_inputs}
\alias{link_inputs}
\title{Symlink Google Drive Data Space to Local Data Directory}
\usage{
link_inputs(cfg_path)
}
\arguments{
\item{cfg_path}{Character. Path to the configuration file containing the Google Drive data space path.}
}
\value{
Logical. Returns TRUE if the linking is successful; otherwise, stops with an error.
}
\description{
This function creates a symbolic link from a Google Drive data space (specified in a config file)
to a local "data" directory. It checks for the existence of the Google Drive path and the local
data directory, creates the local directory if it does not exist, and then creates the symlink.
}
\examples{
\dontrun{
link_inputs("config.yml")
}

}
