% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MadagascarWeatherStations-package.R
\docType{package}
\name{MadagascarWeatherStations-package}
\alias{MadagascarWeatherStations}
\alias{MadagascarWeatherStations-package}
\title{MadagascarWeatherStations: Madagascar Weather Stations}
\description{
A data pipeline and distribution package for our Madagascar weather stations data.
}
\author{
\strong{Maintainer}: Tinashe Michael Tapera \email{ttapera@hsph.harvard.edu}

}
\keyword{internal}
